Refactor the Kotlin Count Functions
Remove the private utility method and use function composition instead
