+++ src/main/kotlin/org/mybatis/dynamic/sql/util/kotlin/KotlinModelBuilderFunctions.kt
@@ -23,52 +23,53 @@ import org.mybatis.dynamic.sql.insert.GeneralInsertDSL
 import org.mybatis.dynamic.sql.insert.InsertDSL
 import org.mybatis.dynamic.sql.insert.MultiRowInsertDSL
 
-fun count(column: BasicColumn, completer: CountColumnCompleter) =
-    completer(KotlinCountColumnBuilder(SqlBuilder.countColumn(column))).build()
+object KotlinModelBuilderFunctions {
+    fun count(column: BasicColumn, completer: CountCompleter) =
+        completer(KotlinCountBuilder(SqlBuilder.countColumn(column))).build()
 
-fun countDistinct(column: BasicColumn, completer: CountColumnCompleter) =
-    completer(KotlinCountColumnBuilder(SqlBuilder.countDistinctColumn(column))).build()
+    fun countDistinct(column: BasicColumn, completer: CountCompleter) =
+        completer(KotlinCountBuilder(SqlBuilder.countDistinctColumn(column))).build()
 
-fun count(start: KotlinCountBuilder, completer: CountCompleter) =
-    completer(start).build()
+    fun countFrom(table: SqlTable, completer: CountCompleter) =
+        with(KotlinCountBuilder(SqlBuilder.countColumn(SqlBuilder.constant<Long>("*")))) {
+            completer(from(table)).build()
+        }
 
-fun countFrom(table: SqlTable, completer: CountCompleter) =
-    count(KotlinCountBuilder(SqlBuilder.countFrom(table)), completer)
+    fun deleteFrom(table: SqlTable, completer: DeleteCompleter) =
+        completer(KotlinDeleteBuilder(SqlBuilder.deleteFrom(table))).build()
 
-fun deleteFrom(table: SqlTable, completer: DeleteCompleter) =
-    completer(KotlinDeleteBuilder(SqlBuilder.deleteFrom(table))).build()
+    fun insertInto(table: SqlTable, completer: GeneralInsertCompleter) =
+        completer(GeneralInsertDSL.insertInto(table)).build()
 
-fun insertInto(table: SqlTable, completer: GeneralInsertCompleter) =
-    completer(GeneralInsertDSL.insertInto(table)).build()
+    fun insertSelect(table: SqlTable, completer: InsertSelectCompleter) =
+        with(completer(KotlinInsertSelectSubQueryBuilder())) {
+            SqlBuilder.insertInto(table)
+                .withColumnList(columnList)
+                .withSelectStatement(selectBuilder)
+                .build()
+        }
 
-fun insertSelect(table: SqlTable, completer: InsertSelectCompleter) =
-    with(completer(KotlinInsertSelectSubQueryBuilder())) {
-        SqlBuilder.insertInto(table)
-            .withColumnList(columnList)
-            .withSelectStatement(selectBuilder)
-            .build()
-    }
+    fun <T> BatchInsertDSL.IntoGatherer<T>.into(table: SqlTable, completer: BatchInsertCompleter<T>) =
+        completer(into(table)).build()
 
-fun <T> BatchInsertDSL.IntoGatherer<T>.into(table: SqlTable, completer: BatchInsertCompleter<T>) =
-    completer(into(table)).build()
+    fun <T> InsertDSL.IntoGatherer<T>.into(table: SqlTable, completer: InsertCompleter<T>) =
+        completer(into(table)).build()
 
-fun <T> InsertDSL.IntoGatherer<T>.into(table: SqlTable, completer: InsertCompleter<T>) =
-    completer(into(table)).build()
+    fun <T> MultiRowInsertDSL.IntoGatherer<T>.into(table: SqlTable, completer: MultiRowInsertCompleter<T>) =
+        completer(into(table)).build()
 
-fun <T> MultiRowInsertDSL.IntoGatherer<T>.into(table: SqlTable, completer: MultiRowInsertCompleter<T>) =
-    completer(into(table)).build()
+    fun select(vararg columns: BasicColumn, completer: SelectCompleter) =
+        select(columns.asList(), completer)
 
-fun select(vararg columns: BasicColumn, completer: SelectCompleter) =
-    select(columns.asList(), completer)
+    fun select(columns: List<BasicColumn>, completer: SelectCompleter) =
+        completer(KotlinSelectBuilder(SqlBuilder.select(columns))).build()
 
-fun select(columns: List<BasicColumn>, completer: SelectCompleter) =
-    completer(KotlinSelectBuilder(SqlBuilder.select(columns))).build()
+    fun selectDistinct(vararg columns: BasicColumn, completer: SelectCompleter) =
+        selectDistinct(columns.asList(), completer)
 
-fun selectDistinct(vararg columns: BasicColumn, completer: SelectCompleter) =
-    selectDistinct(columns.asList(), completer)
+    fun selectDistinct(columns: List<BasicColumn>, completer: SelectCompleter) =
+        completer(KotlinSelectBuilder(SqlBuilder.selectDistinct(columns))).build()
 
-fun selectDistinct(columns: List<BasicColumn>, completer: SelectCompleter) =
-    completer(KotlinSelectBuilder(SqlBuilder.selectDistinct(columns))).build()
-
-fun update(table: SqlTable, completer: UpdateCompleter) =
-    completer(KotlinUpdateBuilder(SqlBuilder.update(table))).build()
+    fun update(table: SqlTable, completer: UpdateCompleter) =
+        completer(KotlinUpdateBuilder(SqlBuilder.update(table))).build()
+}
