+++ src/main/kotlin/org/mybatis/dynamic/sql/util/kotlin/spring/NamedParameterJdbcTemplateExtensions.kt
@@ -29,7 +29,6 @@ import org.mybatis.dynamic.sql.insert.render.MultiRowInsertStatementProvider
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider
 import org.mybatis.dynamic.sql.update.render.UpdateStatementProvider
 import org.mybatis.dynamic.sql.util.kotlin.BatchInsertCompleter
-import org.mybatis.dynamic.sql.util.kotlin.CountColumnCompleter
 import org.mybatis.dynamic.sql.util.kotlin.CountCompleter
 import org.mybatis.dynamic.sql.util.kotlin.DeleteCompleter
 import org.mybatis.dynamic.sql.util.kotlin.GeneralInsertCompleter
@@ -50,10 +49,10 @@ import java.sql.ResultSet
 fun NamedParameterJdbcTemplate.count(selectStatement: SelectStatementProvider) =
     queryForObject(selectStatement.selectStatement, selectStatement.parameters, Long::class.java)!!
 
-fun NamedParameterJdbcTemplate.count(column: BasicColumn, completer: CountColumnCompleter) =
+fun NamedParameterJdbcTemplate.count(column: BasicColumn, completer: CountCompleter) =
     count(org.mybatis.dynamic.sql.util.kotlin.spring.count(column, completer))
 
-fun NamedParameterJdbcTemplate.countDistinct(column: BasicColumn, completer: CountColumnCompleter) =
+fun NamedParameterJdbcTemplate.countDistinct(column: BasicColumn, completer: CountCompleter) =
     count(org.mybatis.dynamic.sql.util.kotlin.spring.countDistinct(column, completer))
 
 fun NamedParameterJdbcTemplate.countFrom(table: SqlTable, completer: CountCompleter) =
